Trajectories part2 completed
I following thought the phone the video is much cumbersome, not only is it hard to make it repeat the part I need, but being small it is jsut annoying.
aside from that frustration, I havent had time to check all the code I wrote, thou I understand what the end result is.
If I were to make this from scratch I wouldnt be ready, more practice is needed.